mas cambios de validaciones sobre comentarios del profe

+++ videojuegos-frontend/src/components/Juego.jsx
@@ -10,6 +10,8 @@ const Juego = ({ _id, id, titulo, descripcion, anioLanzamiento, imageUrl }) => {
   const [form, setForm] = useState({ titulo, descripcion, anioLanzamiento });
 
   const handleEliminar = async () => {
+    const confirmar = window.confirm("¿Estas seguro que queres eliminar este juego?");
+    if (!confirmar) return;
     try {
       await eliminarJuegoApi(idReal);
       dispatch(eliminarJuegoRedux(idReal));
@@ -29,20 +31,20 @@ const Juego = ({ _id, id, titulo, descripcion, anioLanzamiento, imageUrl }) => {
   };
 
   if (editando) {
-  return (
-    <div className="card mb-3">
-      <div className="card-body">
-        <div className="d-flex gap-2 flex-wrap">
-          <input className="form-control" style={{width: "150px"}} value={form.titulo} onChange={(e) => setForm({ ...form, titulo: e.target.value })} placeholder="Título" />
-          <input className="form-control" style={{width: "200px"}} value={form.descripcion} onChange={(e) => setForm({ ...form, descripcion: e.target.value })} placeholder="Descripción" />
-          <input className="form-control" style={{width: "100px"}} type="number" value={form.anioLanzamiento} onChange={(e) => setForm({ ...form, anioLanzamiento: e.target.value })} placeholder="Año" />
-          <button className="btn btn-success" onClick={handleGuardar}>Guardar</button>
-          <button className="btn btn-secondary" onClick={() => setEditando(false)}>Cancelar</button>
+    return (
+      <div className="card mb-3">
+        <div className="card-body">
+          <div className="d-flex gap-2 flex-wrap">
+            <input className="form-control" style={{ width: "150px" }} value={form.titulo} onChange={(e) => setForm({ ...form, titulo: e.target.value })} placeholder="Título" />
+            <input className="form-control" style={{ width: "200px" }} value={form.descripcion} onChange={(e) => setForm({ ...form, descripcion: e.target.value })} placeholder="Descripción" />
+            <input className="form-control" style={{ width: "100px" }} type="number" value={form.anioLanzamiento} onChange={(e) => setForm({ ...form, anioLanzamiento: e.target.value })} placeholder="Año" />
+            <button className="btn btn-success" onClick={handleGuardar}>Guardar</button>
+            <button className="btn btn-secondary" onClick={() => setEditando(false)}>Cancelar</button>
+          </div>
         </div>
       </div>
-    </div>
-  );
-}
+    );
+  }
 
   return (
     <div className="card mb-3">
